move interface to models file
continue moving to global context variable
